move PlayerState class into separate file
